refactoring GraphQL API

+++ resolvers/index.js
@@ -1,3 +1,4 @@
+require("dotenv").config();
 import { ApolloError } from "apollo-server";
 import { UserModel, PostModel, CategoryModel } from "../models";
 import { createPassword, checkPassword, generateToken } from "../utils";
@@ -5,8 +6,10 @@ import { createPassword, checkPassword, generateToken } from "../utils";
 export const resolvers = {
   Query: {
     users: async () => await UserModel.find().sort({ name: 1 }),
-    user: async (_, { name }) => {
-      return await UserModel.find({ name: { $regex: name, $options: "i" } });
+    me: async (_, args, { user }) => {
+      if (!user) throw new Error("You are not authenticated");
+      return await UserModel.findById(user.id);
+      // return await UserModel.find({ name: { $regex: name, $options: "i" } });
     },
     posts: async () =>
       await PostModel.find({ deleted: false }).sort({ updateAt: -1 }),
@@ -23,11 +26,11 @@ export const resolvers = {
     createUser: async (_, { input }) => {
       try {
         const user = await UserModel.findOne({ email: input.email });
-        if (user) throw new ApolloError("User already exist!");
+        if (user) throw new Error("User already exist!");
         input.password = await createPassword(input.password);
         return await UserModel.create(input);
       } catch (error) {
-        console.log('error:' + error)
+        console.log("error:" + error);
       }
     },
     createCategory: async (_, { input }) => {
@@ -35,30 +38,47 @@ export const resolvers = {
       if (category) throw new ApolloError("Category already exist!");
       return await CategoryModel.create(input);
     },
-    createPost: async (_, { input }) => {
-      input.deleted = false;
-      return await PostModel.create(input);
+    createPost: async (_, { input }, { user }) => {
+      try {
+        if (!user)
+          throw new Error("You are not authenticated to post an article");
+        const check_user = await UserModel.findById(user.id);
+        if (check_user) {
+          input.deleted = false;
+          input.author = user.id;
+          const post = await PostModel.create(input);
+          check_user.posts.push(post._id);
+          await UserModel.updateOne(
+            { _id: user.id },
+            { $set: check_user },
+            { new: true }
+          );
+          return post;
+        }
+      } catch (error) {}
     },
-    deletePost: async (_, { id }) => {
+    deletePost: async (_, { id }, { user }) => {
+      if (!user)
+        throw new Error("You are not authenticated to delete this article");
       await PostModel.deleteOne({ _id: id });
       return await PostModel.find().sort({ updateAt: -1 });
     },
-    updatePost: async (_, { id, input }) => {
+    updatePost: async (_, { id, input }, { user }) => {
+      if (!user)
+        throw new Error("You are not authenticated to update this article");
       await PostModel.updateOne({ _id: id }, { $set: input }, { new: true });
       return PostModel.find().sort({ updateAt: -1 });
     },
     login: async (_, { email, password }) => {
       const user = await UserModel.findOne({ email: email });
       if (user) {
-        const isPassword = await checkPassword(user.password, password);
-        if (isPassword) {
-          user.token = generateToken({userID: user._id, userName: user.name});
-          return user;
+        const check_password = await checkPassword(user.password, password);
+        if (check_password) {
+          const token = await generateToken(user._id);
+          return { token, user };
         }
       }
-      return new ApolloError(
-        "Login failed 😭. Please verify your email or password!"
-      );
+      throw new Error("Login failed 😭. Please verify your email or password!");
     },
   },
 };
